ocr-server/requirements.txt
fastapi==0.115.0
uvicorn==0.32.0
python-multipart==0.0.12
paddlepaddle==2.6.2
paddleocr==2.9.1
Pillow==10.4.0
numpy==1.26.4
